# ranking: Optimize sloppy query