@ocanty 2022-11-10
@Strum355 sent me here to look at this masterpiece