@chrismwendt 2022-11-10
"do the smaller loop first" 😄